Create README.md
In this project we used Python libraries like NumPy, pandas, matplotlib and seaborn to analysis data about the Olympics. The data used is from Kaggle titled - 120 years of Olympic history ; and it shows Olympic games athletes and results from Athens 1896 to Rio 2016. This project was inspired by the 2020 Tokyo Olympics. 
			
Findings from this project showed the USA has been a dominated force in these games, it has had the most participants(17847) to date. France and Great Britain were close with 11988 and 11404 participants respectively. The surprising country that showed up in the top 10 was Hungary with 6547 athletes participating over time, this came as a surprise because they are a relatively smaller country.
			
Next graphed the age distribution of athletes who have competed in the Olympics and it came as no shock that 20-30 year old athletes were the most common. This speak to their physical condition at that age, it is the most ideal to compete in these rigorous sports. 
			
Looking at gender, there have been 196594 male athletes who have competed since the start of the games vs the 74522 females. It was refreshing to find that female athlete numbers have been gradually increasing since 1980, and can only expect those numbers to keep increasing thanks to the elimination of social norms that hindered females from competing before. 
			
Lastly there is a high correlation (0.8) between the weight and height of athletes who have earned medals in the Olympics. Again athletes have to be in the best of shape to perform at their optimal best.